+++ modules/manager/controller_setup.go
@@ -102,6 +102,11 @@ func SetupCacheIndexes(ctx context.Context, mgr manager.Manager, cfg Config) err
 		Version:  gatewayv1.GroupVersion.Version,
 		Resource: "tlsroutes",
 	}
+	tcpRouteGVR := schema.GroupVersionResource{
+		Group:    gatewayv1.GroupVersion.Group,
+		Version:  gatewayv1.GroupVersion.Version,
+		Resource: "tcproutes",
+	}
 
 	if cfg.ControlPlaneControllerEnabled || cfg.GatewayControllerEnabled {
 		indexOptions = slices.Concat(indexOptions,
@@ -126,6 +131,13 @@ func SetupCacheIndexes(ctx context.Context, mgr manager.Manager, cfg Config) err
 		}
 		if hasTLSRoute {
 			indexOptions = slices.Concat(indexOptions, index.OptionsForTLSRoute())
+		}
+		hasTCPRoute, err := crdChecker.CRDExists(tcpRouteGVR)
+		if err != nil {
+			return fmt.Errorf("failed to check existence of CRD %s: %w", tcpRouteGVR.String(), err)
+		}
+		if hasTCPRoute {
+			indexOptions = slices.Concat(indexOptions, index.OptionsForTCPRoute())
 		}
 	}